format: decomposition
